refactor(config): move config from internal to pkg
Pro and any other downstream binary that calls app.New(cfg) needs to
construct a config.Config first. Keeping config in internal/ blocked
that. Moved to pkg/config so the Config struct and Load() helper are
part of Core's public surface.

No behavior change. Sensitive fields (JWTSecret) are still only populated
from environment variables — moving the package does not weaken any
protection.
